+++ src/tasks/Deploy.js
@@ -28,7 +28,7 @@ class DeployTask extends ApiTask {
 		const { stack } = this.options;
 		return Object.assign({}, stack.params, {
 			StackName: stack.name,
-			TemplateBody: JSON.stringify(this.inputArtifacts.template),
+			TemplateBody: JSON.stringify(this.input.template),
 			ClientRequestToken: this.taskUUID,
 		});
 	}
@@ -68,6 +68,7 @@ class DeployTask extends ApiTask {
 			if (err) {
 				this.log.error(`${err.code}: ${err.message}`, false);
 				this.log.info(`└─ RequestId: ${chalk.magenta(err.requestId)}`);
+				this.log.stop();
 				process.exit(1);
 			} else {
 				this.log.message(message);